+++ scripts/macros/monsterFeatures/autognome/autognome.js
@@ -1,0 +1,4 @@
+import {malfunction} from './malfunction.js';
+export let autoGnome = {
+    'malfunction': malfunction
+}